WIP: Attempting to fix the import order
