+++ main.py
@@ -30,3 +30,7 @@ async def discord(response: Request): #
 
         context = DiscordContext(member=json.member)
         return json.data.execute(context)
+
+
+if __name__ == '__main__':
+    uvicorn.run(app=app, host="localhost", port=8181)
